+++ lib/grape/dsl/inside_route.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'grape/dsl/headers'
-
 module Grape
   module DSL
     module InsideRoute