+++ lib/grape/json.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
-require 'json'
-
 module Grape
-  if Object.const_defined? :MultiJson
+  if defined?(::MultiJson)
     Json = ::MultiJson
   else
     Json = ::JSON