+++ lib/grape/validations/validators/base.rb
@@ -59,6 +59,7 @@ module Grape
         end
 
         def self.inherited(klass)
+          super
           return if klass.name.blank?
 
           short_validator_name = klass.name.demodulize.underscore.delete_suffix('_validator')