+++ lib/grape/xml.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Grape
-  if Object.const_defined? :MultiXml
+  if defined?(::MultiXml)
     Xml = ::MultiXml
   else
     Xml = ::ActiveSupport::XmlMini